requirements.txt
pandas>=0.24.1
Flask==1.1.1
boto
requests
mysql-connector-python==1.2.3